Move ConnectionManangerBuilder to submodule
Moves the builder to crate::network::connection_manager::builder in an
effort to make connection manager modules more consumable by making the
top-level module file shorter.
